Remove Tag interface
