+++ src/main/java/org/toradocu/extractor/ParamTag.java
@@ -11,7 +11,7 @@ import org.toradocu.util.Checks;
  * expressed by this tag is violated, and the behavior of the method documented by this tag is
  * unspecified.
  */
-public class ParamTag extends AbstractTag {
+public class ParamTag extends Tag {
 
   /** The parameter associated with the param tag */
   private final Parameter parameter;
